requirements.txt
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-jet-reboot==1.3.1
django-rest-authtoken==2.1.4
django-search-admin-autocomplete==0.2.1
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
facebook-sdk==3.1.0
Faker==14.2.0
feedparser==6.0.10
firebase-admin==5.3.0
future==0.18.2
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.48.1
grpcio-status==1.48.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
proto-plus==1.22.1
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-twitter==3.5
python3-openid==3.2.0
pytz==2022.2.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
toposort==1.7
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
